Reviews can be Updated

+++ src/routes/reviewRoute.js
@@ -20,6 +20,23 @@ reviewRoute.post("/review/add/:storeid", userAuth, async (req, res) => {
   }
 });
 
+reviewRoute.put("/review/:id", userAuth, async (req, res) => {
+  const updatedReview = { ...req.body };
+  const foundReview = await review
+    .find({ _id: req.params.id, user_id: req.user._id })
+    .exec();
+
+  try {
+    if (foundReview.length == 0) {
+      throw new Error("No such Review Exists");
+    }
+    await review.findOneAndUpdate({ _id: req.params.id }, updatedReview);
+    res.status(200).send({ msg: "Review Updated" });
+  } catch (error) {
+    res.status(400).send(error.message);
+  }
+});
+
 reviewRoute.get("/reviews", storeAuth, async (req, res) => {
   try {
     await req.store.populate("reviews");
@@ -39,14 +56,21 @@ reviewRoute.get("/myReviews", userAuth, async (req, res) => {
 });
 
 reviewRoute.delete("/review/:id", userAuth, async (req, res) => {
+  const foundReview = await review
+    .find({ _id: req.params.id, user_id: req.user._id })
+    .exec();
+
   try {
+    if (foundReview.length == 0) {
+      throw new Error("No such Review Exists");
+    }
     const reviews = await review.findOneAndDelete({
       _id: req.params.id,
       user_id: req.user._id,
     });
     res.status(200).send(reviews);
   } catch (error) {
-    res.status(400).send(error);
+    res.status(400).send(error.message);
   }
 });
